Add Evaluation Modules
Still needs a bit of work, but built the evaluation modules which will be called by evaluation handler after that modules receives a response from an llm or is elsewise directed. Included is the perplexity algorithm, bleu and rouge. I have tried to build it to follow the factory architecture to allow for multiprocessing in the future.